+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/Live/LiveNow/Timer.tsx
@@ -1,0 +1,39 @@
+import { Text } from '@codesandbox/components';
+import React, { FunctionComponent, useState } from 'react';
+import useInterval from 'use-interval';
+import { useOvermind } from 'app/overmind';
+
+const pad = (number: number) => {
+  if (`${number}`.length === 1) {
+    return `0${number}`;
+  }
+
+  return `${number}`;
+};
+
+export const Timer: FunctionComponent = () => {
+  const {
+    state: {
+      live: {
+        roomInfo: { startTime },
+      },
+    },
+  } = useOvermind();
+  const [since, setSince] = useState(Date.now() - startTime);
+
+  useInterval(() => {
+    setSince(Date.now() - startTime);
+  }, 1000);
+
+  const hours = Math.floor(since / 1000 / 60 / 60);
+  const minutes = Math.floor((since - hours * 1000 * 60 * 60) / 1000 / 60);
+  const seconds = Math.floor(
+    (since - hours * 1000 * 60 * 60 - minutes * 1000 * 60) / 1000
+  );
+
+  const text = `${hours > 0 ? pad(hours) + ':' : ''}${pad(minutes)}:${pad(
+    seconds
+  )}`;
+
+  return <Text variant="danger">{text}</Text>;
+};
