+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/Live/NotLoggedIn.tsx
@@ -1,15 +1,15 @@
-import React from 'react';
-
 import {
-  Element,
+  Button,
   Collapsible,
+  Element,
   Stack,
   Text,
-  Button,
 } from '@codesandbox/components';
+import React, { FunctionComponent } from 'react';
+
 import { useOvermind } from 'app/overmind';
 
-export const NotLoggedIn = () => {
+export const NotLoggedIn: FunctionComponent = () => {
   const {
     actions: { signInClicked },
   } = useOvermind();
@@ -17,19 +17,22 @@ export const NotLoggedIn = () => {
   return (
     <Collapsible title="Live" defaultOpen>
       <Element paddingX={2}>
-        <Text block weight="medium" marginBottom={2}>
+        <Text block marginBottom={2} weight="medium">
           Collaborate in real-time
         </Text>
+
         <Stack direction="vertical" gap={2} marginBottom={6}>
-          <Text size={2} variant="muted" block>
+          <Text block size={2} variant="muted">
             You need to be signed in to open a live session to collaborate with
             others in real time.
           </Text>
-          <Text size={2} variant="muted" block>
+
+          <Text block size={2} variant="muted">
             Sign in to live share this sandbox!
           </Text>
         </Stack>
-        <Button variant="primary" onClick={() => signInClicked()}>
+
+        <Button onClick={() => signInClicked()} variant="primary">
           Sign in with GitHub
         </Button>
       </Element>
