+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/Live/index.tsx
@@ -1,8 +1,10 @@
 import React, { FunctionComponent } from 'react';
+
 import { useOvermind } from 'app/overmind';
-import { NotLoggedIn } from './NotLoggedIn';
+
 import { LiveNow } from './LiveNow';
 import { NotLive } from './NotLive';
+import { NotLoggedIn } from './NotLoggedIn';
 
 export const Live: FunctionComponent = () => {
   const {
@@ -12,6 +14,9 @@ export const Live: FunctionComponent = () => {
     },
   } = useOvermind();
 
-  if (!isLoggedIn) return <NotLoggedIn />;
+  if (!isLoggedIn) {
+    return <NotLoggedIn />;
+  }
+
   return isLive ? <LiveNow /> : <NotLive />;
 };
